% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_diabetes}
\alias{df_diabetes}
\title{Diabetes data set}
\format{A data frame with 442 observations
\describe{
  \item{y}{continuous-valued}
  \item{x}{features, 10 in total.}
}}
\source{
\url{http://web.stanford.edu/~hastie/Papers/LARS/}
}
\usage{
df_diabetes
}
\description{
Diabetes data set
}
\keyword{datasets}
