% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsel_explore.R
\name{varsel_explore}
\alias{varsel_explore}
\title{Interactive variable selection using glmproj}
\usage{
varsel_explore(fit_cv, nv = min(12, length(fit_cv$varsel$vind)))
}
\arguments{
\item{fit_cv}{An object returned by \link[=cv_varsel]{cv_varsel}.}

\item{nv}{Maximum number of variables in the submodel. Defaults to
\code{min(12, length(fit_cv$varsel$vind))}.}
}
\description{
Visualise the results of the variable selection using shiny.
}
