% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_diabetes}
\alias{df_diabetes}
\title{Diabetes data set}
\format{A data frame with 442 observations and the following variables:
\describe{
  \item{age}{Age, normalized.}
  \item{sex}{Sex.}
  \item{bmi}{Bmi, normalized.}
  \item{map}{Map, normalized.}
  \item{tc}{Tc, normalized.}
  \item{ldl}{Ldl, normalized.}
  \item{hdl}{Hdl, normalized.}
  \item{tch}{Tch, normalized.}
  \item{ltg}{Ltg, normalized.}
  \item{glu}{Glu, normalized.}
  \item{y}{Target variable.}
}}
\source{
\url{http://web.stanford.edu/~hastie/Papers/LARS/}
}
\usage{
df_diabetes
}
\description{
Diabetes data set
}
\keyword{datasets}
